Update __repr__() method models.py
I chose this additional piece of code (which.. "makes an attempt to return a string that would yield an object with the same value when passed to eval()") because models and test_crud are important in order to have the testing phase of this application work correctly. In the testing phase it is important to have access to printable versions of objects to see and determine what they contain. It is important to be able to easily access what an object contains easily via printing so that at any point of the complex process of testing, the developer can refer to and “evaluate” any object in the correct way.